+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # 0.2.30
 
-* Added formatting control for `AbslStringify` externder using struct and ADL method `MboTypesExtendStringifyOptions`.
+* Added formatting control for `AbslStringify` externder using `AbslStringifyOptions`.
 * Added `AbslStringify` ADL API extension entry points `MboTypesExtendStringifyOptions` and `MboTypesExtendStringifyOptions`.
-* Added function `WithFieldNames` a format control adapter for `AbslStringify`.
-* Added field name support for non literal typesin Clang.
+* Added function `extender::WithFieldNames` a format control adapter for `AbslStringify`.
+* Added field name support for non literal types in Clang.
 
 # 0.2.29